"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ cspp2-assignments/m2/Assignment-3/Assignment-3/Solution.java
@@ -20,7 +20,7 @@ public final class Solution {
      * @return     { description_of_the_return_value }
      */
 
-    static int power(final int base, final int exponent) {
+    static long power(final int base, final int exponent) {
         /**
          * { var_description }.
          */
@@ -40,7 +40,7 @@ public final class Solution {
         Scanner s = new Scanner(System.in);
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result = power(base, exponent);
+        long result = power(base, exponent);
         System.out.println(result);
 
     }
